// src/models/mycard.models.ts
export class MyCard {
    id: string;
    object: string;
    address_city: string;
    address_country: string;
    address_line1: string;
    address_line1_check: string;
    address_line2: string;
    address_state: string;
    address_zip: string;
    address_zip_check: string;
    brand: string;
    country: string;
    customer: string;
    cvc_check: string;
    dynamic_last4: string;
    exp_month: string;
    exp_year: string;
    fingerprint: string;
    funding: string;
    last4: string;
    metadata: {
        title: string;
        email: string;
    };
    name: string;
    tokenization_method: string;
}